v1.2/phase 2 type conformance (#30)
* docs(02): capture phase context

* docs(state): record phase 2 context session

* docs(02): create phase plan

* docs(02): create phase plan

* refactor(02-01): retype modify_order money params to Optional[Decimal]

- DEC-01/D-03: modify_order new_price/new_quantity -> Optional[Decimal] at
  OrderHandler facade + OrderManager (no float-for-money at domain boundary)
- Update NumPy docstrings (float -> Decimal) at both layers
- Retain defensive to_money() coercion at order_manager.py:1136-1138
- D-05 minimal: annotations + docstrings only; cancel_order (no money params)
  untouched; no LIFE-01 pre-build

* feat(02-03): add CorrelationId NewType + idgen.generate_correlation_id

- Add CorrelationId = NewType('CorrelationId', uuid.UUID) as 10th core/ids.py alias + __all__ entry (D-02)
- Add IDGenerator.generate_correlation_id(self) -> uuid.UUID mirroring siblings (D-01/D-02)
- Update module docstring Nine -> Ten alias count

* fix(02-02): carry _min/_max_order_size as Decimal end-to-end (DEC-02/D-06)

- drop float() wraps at simulated.py init (:99-100) and update_config (:605-606)
- validate_order comparisons (:388,390) now run Decimal-vs-Decimal
- mirror E2E seam (conftest.py:331-332): drop float() re-derivation
- correct conftest :323 comment block (caches are Decimal, not floats)

* test(02-01): pass Decimal to modify_order at the test boundary (D-04)

- Convert the three in-repo float modify_order callers to Decimal("...")
  so the money boundary is float-free in practice, not only in annotation
- new_price=28.0 -> Decimal("28.0"); new_quantity=2.0 -> Decimal("2.0")
- Behavior unchanged: assertions already expected the Decimal result
- test_order.py (Order entity API) untouched

* test(02-02): add below-minimum REFUSED branch (Decimal-vs-Decimal) — D-08

- closes symmetric < _min_order_size branch (release_refused covers only > _max)
- configures limits via update_config(Decimal literals) to avoid setattr-bypass float
- asserts isinstance(_min_order_size, Decimal) as DEC-02 regression lock
- asserts is_valid False + 'below minimum' message (correct REFUSED, not a TypeError)

* feat(02-03): mint correlation id from idgen, retype event field, drop dead uuid import

- _generate_correlation_id returns CorrelationId(idgen.generate_correlation_id()); ph_ prefix dropped (D-01/D-02)
- Retype ErrorEvent.correlation_id str | None -> CorrelationId | None + docstring (inherited by PortfolioErrorEvent)
- Retype _publish_error_event correlation_id param + _operation_context generator -> CorrelationId
- Remove now-dead 'import uuid' from portfolio_handler.py (Phase-1 D-05 touched-path cleanup)

* docs(02-02): reframe DEC-02 to float-for-money consistency (D-07 delta)

- drop the 'latent Decimal < float TypeError' misdiagnosis from DEC-02
- reframe as float-for-money consistency at the exchange size-limit boundary
- note the symmetric below-minimum REFUSED branch regression-cover (D-08)
- retain [W2-10] traceability tag
- (ROADMAP.md SC-2 reframe + STATE.md D-07 delta log deferred to orchestrator —
   worktree directive forbids executor edits to ROADMAP.md/STATE.md; wording
   captured in 02-02-SUMMARY.md for the central update)

* test(02-03): assert correlation id is a uuid.UUID instead of ph_ prefix

- Swap .startswith('ph_') for isinstance(id, uuid.UUID); keep uniqueness check
- Reflects idgen UUIDv7 correlation id (DEC-03 single-ID-scheme conformance)

* docs(02-03): complete locked-decision-conformance DEC-03 plan

* docs(02-01): complete DEC-01 Optional[Decimal] money-API plan

- SUMMARY: 3 tasks, golden byte-exact / 58 e2e / mypy strict clean
- deferred-items: log DEF-02-01 (cross-agent 02-03 test failure, out of scope)

* docs(02-03): append self-check result to SUMMARY

* docs(02-02): complete decimal-order-size plan summary

* docs(phase-02): update tracking after wave 1 (DEC-01/02/03 complete; D-07 reframe)

* docs(phase-02): complete phase execution

* docs(02): add code review report

* docs(02): mark DEF-02-01 resolved (closed by 57ad3df)

* fix(02): WR-01 make JSON renderer UUID-safe for correlation_id error logging

* fix(02): IN-01 float size limits in get_exchange_info serialization dict

* fix(02): IN-02 update stale float comment to Decimal in modify_order

* docs(02): add code review fix report (WR-01 + 2 info fixed; 2 latent info deferred)

* docs(02): log DEF-02-02/DEF-02-03 residual review items (DEF-02-03 -> Phase 4 TYPE)

* docs(req): extend TYPE-02 with order-API order_id/portfolio_id retype (DEF-02-03 -> Phase 4)

+++ itrader/order_handler/order_manager.py
@@ -1084,7 +1084,7 @@ class OrderManager:
 			error_details=reason,
 			operation_type=operation_type)
 
-	def modify_order(self, order_id: int, new_price: Optional[float] = None, new_quantity: Optional[float] = None,
+	def modify_order(self, order_id: int, new_price: Optional[Decimal] = None, new_quantity: Optional[Decimal] = None,
 	                portfolio_id: Optional[int] = None, reason: str = "user modification") -> OperationResult:
 		"""
 		Modify an existing order and generate OrderEvent.
@@ -1093,15 +1093,15 @@ class OrderManager:
 		----------
 		order_id : int
 			The ID of the order to modify
-		new_price : float, optional
+		new_price : Decimal, optional
 			New price for the order
-		new_quantity : float, optional
+		new_quantity : Decimal, optional
 			New quantity for the order
 		portfolio_id : int, optional
 			Portfolio ID for faster lookup
 		reason : str, optional
 			Reason for the modification
-			
+
 		Returns
 		-------
 		OperationResult
@@ -1130,8 +1130,8 @@ class OrderManager:
 						operation_type="modify_order"
 					)
 			
-			# Apply the modification. Order money is Decimal (M2a); coerce the
-			# float modify args at this boundary.
+			# Apply the modification. Order money is Decimal (M2a); normalize the
+			# Decimal modify args through the money entry point at this boundary.
 			success = order.modify_order(
 				to_money(new_price) if new_price is not None else None,
 				to_money(new_quantity) if new_quantity is not None else None,
